Fixed some of the comments from Felipe

+++ src/main/java/com/singular/renting/domain/PriceType.java
@@ -1,16 +1,18 @@
 package com.singular.renting.domain;
 
+import java.math.BigDecimal;
+
 public enum PriceType {
-    PREMIUM (3.0f),
-    BASIC   (1.0f);
+    PREMIUM (BigDecimal.valueOf(3.0)),
+    BASIC   (BigDecimal.valueOf(1.0));
 
-    private final Float value;
+    private final BigDecimal value;
 
-    PriceType(Float value) {
+    PriceType(BigDecimal value) {
         this.value = value;
     }
 
-    public Float getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 }
